modified viewer for slice cell types so it can visualize pml cells

+++ src/wavesolver/MAC_Grid.h
@@ -276,7 +276,7 @@ class MAC_Grid
         void classifyCellsDynamic_FAST(MATRIX &pFull, MATRIX (&p)[3], FloatArray &pGCFull, FloatArray (&pGC)[3], MATRIX (&v)[3], const bool &useBoundary, const bool &verbose=false);
         void classifyCellsDynamicAABB(const bool &useBoundary, MATRIX &p, const bool &verbose=false);
         void ComputeGhostCellSolveResidual(const FloatArray &p, REAL &minResidual, REAL &maxResidual, REAL &maxOffDiagonalEntry); 
-        int PressureCellType(const int &idx);
+        REAL PressureCellType(const int &idx);
         void ResetCellHistory(const bool &valid); 
         void GetCell(const int &cellIndex, MATRIX const (&pDirectional)[3], const MATRIX &pFull, const MATRIX (&v)[3], Cell &cell) const;
